requirements.txt
# ============ 核心 ============
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
PyYAML>=6.0
requests>=2.28.0

# ============ 数据存储 ============
clickhouse-driver>=0.2.6
redis>=4.5.0

# ============ 机器学习 ============
lightgbm>=4.0.0
torch>=2.0.0
onnxruntime-gpu>=1.16.0    # GPU推理; CPU版用 onnxruntime
scikit-learn>=1.3.0

# ============ NLP ============
transformers>=4.30.0
beautifulsoup4>=4.12.0

# ============ QMT ============
# xtquant 需从券商获取, 不在PyPI

# ============ 监控 ============
prometheus-client>=0.17.0

# ============ 工具 ============
schedule>=1.2.0
loguru>=0.7.0               # 可选, 增强日志

# ============ GPU加速 (可选) ============
# cupy-cuda12x>=12.0.0

